+++ simvex/src/main/java/dochigosum/simvex/domain/drawing/presentation/dto/request/MemoCreateRequest.java
@@ -1,0 +1,6 @@
+package dochigosum.simvex.domain.drawing.presentation.dto.request;
+
+public record MemoCreateRequest(
+        Long drawingId,
+        String detail
+) {}
