+++ simvex/src/main/java/dochigosum/simvex/domain/drawing/presentation/dto/response/MemoCreatedResponse.java
@@ -1,0 +1,5 @@
+package dochigosum.simvex.domain.drawing.presentation.dto.response;
+
+public record MemoCreatedResponse (
+        Long created
+) {}
