+++ simvex/src/main/java/dochigosum/simvex/domain/drawing/repository/MemoRepository.java
@@ -1,6 +1,6 @@
-package dochigosum.simvex.domain.memo.repository;
+package dochigosum.simvex.domain.drawing.repository;
 
-import dochigosum.simvex.domain.memo.entity.Memo;
+import dochigosum.simvex.domain.drawing.entity.Memo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemoRepository extends JpaRepository<Memo, Long> {
