--- simvex/src/main/java/dochigosum/simvex/domain/memo/presentation/dto/MemoCreateRequest.java
@@ -1,9 +1,0 @@
-package dochigosum.simvex.domain.memo.presentation.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public record MemoCreateRequest(
-        Long drawingId,
-        String detail
-) {}
