--- simvex/src/main/java/dochigosum/simvex/domain/memo/presentation/dto/MemoCreatedResponse.java
@@ -1,8 +1,0 @@
-package dochigosum.simvex.domain.memo.presentation.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-public record MemoCreatedResponse (
-    Long created
-) {}
